Added stuff for card payments

+++ Sycade.BunqApi/Endpoints/PaymentEndpoint.cs
@@ -14,21 +14,21 @@ namespace Sycade.BunqApi.Endpoints
             : base(apiClient) { }
 
 
-        public async Task<Id> CreateAsync(int fromAccountId, Alias to, Amount amount, string description, Session session)
+        public async Task<Id> CreateAsync(long fromAccountId, Alias to, Amount amount, string description, Session session)
         {
             var request = new CreatePaymentRequest(amount, to, description);
 
             return await ApiClient.DoSignedApiRequestAsync<Id>(HttpMethod.Post, $"user/{session.User.Id}/monetary-account/{fromAccountId}/payment", session.Token, request);
         }
 
-        public async Task<Payment[]> GetAllAsync(int monetaryAccountId, Session session)
+        public async Task<Payment[]> GetAllAsync(long monetaryAccountId, Session session)
         {
             var entities = await ApiClient.DoSignedApiRequestAsync(HttpMethod.Get, $"user/{session.User.Id}/monetary-account/{monetaryAccountId}/payment", session.Token);
 
             return entities.Cast<Payment>().ToArray();
         }
 
-        public async Task<Payment> GetByIdAsync(int paymentId, int monetaryAccountId, Session session)
+        public async Task<Payment> GetByIdAsync(long paymentId, long monetaryAccountId, Session session)
         {
             return await ApiClient.DoSignedApiRequestAsync<Payment>(HttpMethod.Get, $"user/{session.User.Id}/monetary-account/{monetaryAccountId}/payment/{paymentId}", session.Token);
         }
